renaming temporalcommuntiies to communitymeasures
