+++ teneto/communitymeasures/allegiance.py
