+++ teneto/communitymeasures/flexibility.py
