+++ test/communitymeasures/test_persistence.py
@@ -1,5 +1,5 @@
 import numpy as np
-from teneto.temporalcommunity import persistence
+from teneto.communitymeasures import persistence
 
 def test_persistence():
     temporalcommunities = np.array([[0, 0, 0, 0], [0, 0, 1, 1], [1, 1, 2, 1]])
